Move TreeLike impl to concrete module
When we implemented the `TreeLike` trait for the concrete `Policy` we
put it in the `iter` module, it doesn't really live there, better to put
the impl in the same place as the definition of the implementor.

Move the impl of `TreeLike` for `concrete::Policy` to the `concrete`
module.
